Account: extract `CancelVerification` flow helper
