+++ components/brave_account/BUILD.gn
@@ -15,6 +15,8 @@ source_set("brave_account") {
     "brave_account_service.h",
     "brave_account_ui_base.h",
     "brave_account_utils.h",
+    "flows/cancel_verification.cc",
+    "flows/cancel_verification.h",
     "flows/change_password.cc",
     "flows/change_password.h",
     "flows/get_service_token.cc",
@@ -132,6 +134,7 @@ source_set("unit_tests") {
   sources = [
     "brave_account_service_test.h",
     "brave_account_service_unittest.cc",
+    "flows/cancel_verification_unittest.cc",
     "flows/change_password_unittest.cc",
     "flows/get_service_token_unittest.cc",
     "flows/log_out_unittest.cc",
