+++ components/brave_account/brave_account_service_unittest.cc
@@ -191,52 +191,4 @@ INSTANTIATE_TEST_SUITE_P(
                     AuthenticationObserverLoggedInToLoggedInEmailChange()),
     BraveAccountServiceAuthenticationObserverTest::kNameGenerator);
 
-struct CancelVerificationTestCase {
-  static void Run(const CancelVerificationTestCase& test_case,
-                  PrefService& pref_service,
-                  mojo::Remote<mojom::Authentication>& authentication) {
-    AccountStatePrefs account_state_prefs(pref_service);
-    account_state_prefs.AddVerification(
-        EncryptedVerificationToken(),
-        mojom::VerificationIntent::NewLoggedOutIntent(
-            mojom::LoggedOutVerificationIntent::kRegistration));
-    authentication->CancelVerification(
-        mojom::VerificationIntent::NewLoggedOutIntent(
-            mojom::LoggedOutVerificationIntent::kRegistration));
-    authentication.FlushForTesting();
-    const auto state = account_state_prefs.GetAccountState();
-    ASSERT_TRUE(state->is_logged_out());
-    EXPECT_FALSE(state->get_logged_out()->verification);
-  }
-
-  std::string test_name;
-};
-
-namespace {
-
-const CancelVerificationTestCase*
-CancelVerificationVerificationTokenNonEmpty() {
-  static const base::NoDestructor<CancelVerificationTestCase>
-      kCancelVerificationVerificationTokenNonEmpty({
-          .test_name = "cancel_registration_verification_token_non_empty",
-      });
-  return kCancelVerificationVerificationTokenNonEmpty.get();
-}
-
-using BraveAccountServiceCancelVerificationTest =
-    BraveAccountServiceTest<CancelVerificationTestCase>;
-
-}  // namespace
-
-TEST_P(BraveAccountServiceCancelVerificationTest,
-       HandlesCancelVerificationOutcomes) {
-  RunTestCase();
-}
-
-INSTANTIATE_TEST_SUITE_P(
-    BraveAccountServiceTests,
-    BraveAccountServiceCancelVerificationTest,
-    testing::Values(CancelVerificationVerificationTokenNonEmpty()),
-    BraveAccountServiceCancelVerificationTest::kNameGenerator);
-
 }  // namespace brave_account
